Fix missing post mortem template content.
Terraform requires values from the server to match the values in configuration.
Rootly returns template content marked up with HTML, which doesn't match the value in configuration.

To get around this, a DiffSuppressFunc is provided in the Terraform schema which returns true if the length of the old value is greater than 0.
The first time the configuration is diffed, the old value will be an empty string, so Terraform will detect the initial value set in configuration and use it to create the resource.
When Terraform diffs the values from the server's response, the length of the old value will be greater than 0, so it will ignore the new value.